fix: what the first compile found
Four, all of them things only a compiler sees. My own revert to four corners left
a stale identifier and a comment still describing eight slots. ktlint wants an
empty line before a declaration that carries a comment. MacroWindow.Move hid
Control's own Move event, which warnings-as-errors makes fatal rather than
merely shadowed. And the Font overload taking a family name wants a float, so an
int straight from Layout.Scale sent the call to the FontFamily overload instead
and complained about the argument before it.

+++ windows/src/Edgepad/Controls/LevelOverlay.cs
@@ -247,8 +247,8 @@ internal sealed class LevelOverlay : IDisposable
                 laidOutAt = dpi;
                 labelFont.Dispose();
                 valueFont.Dispose();
-                labelFont = new Font(Face, Layout.Scale(LabelSize, dpi), FontStyle.Regular, GraphicsUnit.Pixel);
-                valueFont = new Font(Face, Layout.Scale(ValueSize, dpi), FontStyle.Regular, GraphicsUnit.Pixel);
+                labelFont = new Font(Face, (float)Layout.Scale(LabelSize, dpi), FontStyle.Regular, GraphicsUnit.Pixel);
+                valueFont = new Font(Face, (float)Layout.Scale(ValueSize, dpi), FontStyle.Regular, GraphicsUnit.Pixel);
             }
 
             if (Bounds == where)
